feat : 클릭 시 레벨별 슬라임 이미지 추가

+++ Assets/02.Scripts/Ingame/Manager/GameManager.cs
@@ -5,11 +5,19 @@ public class GameManager : MonoBehaviour
 {
     public static GameManager Instance;
 
-    public int AutoPoint;
-    public int ManualPoint;
+    public double AutoPoint;
+    public double ManualPoint;
 
-    private int _point;
-    public int Point
+    // int(21억) long(경) bigInteger(숫자를 쪼개서 계산하기 때문에 매우 느림)
+    // float(10^38), double(10^900), decimal (부동소수점) 같은 자료형 크기 대비 범위가 엄청나게 큼
+    // float의 정밀도 7자리
+    // double의 정밀도 15~16자리
+    // decimal의 정밀도 28~29자리
+    // 숫자의 레벨이 10조단위~ 경단위로 넘어간다면 1, 10과 같은 엄청 작은 숫자는 신경도 쓰지 않음.
+    // double로 게임 포인트를 작성하면 1단위의 값은 정확할지 모르지만 계산이 빠름
+    // 만약 1단위의 계산도 정확하게 하고 싶다면 bigInteger를 사용
+    private double _point;
+    public double Point
     {
         get => _point;
         set
@@ -19,7 +27,7 @@ public class GameManager : MonoBehaviour
         }
     }
 
-    public event Action<int> OnPointChanged;
+    public event Action<double> OnPointChanged;
 
     private void Awake()
     {
@@ -33,7 +41,7 @@ public class GameManager : MonoBehaviour
         }
     }
 
-    public void AddPoint(int amount)
+    public void AddPoint(double amount)
     {
         Point += amount;
     }
